+++ src/main/java/park/brothers/runwith_back/domain/Login/dto/Request/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package park.brothers.runwith_back.domain.Login.dto;
+package park.brothers.runwith_back.domain.Login.dto.Request;
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
